[FEATURE] Redirection (Phase 1: redirect when missing/untranslatable argument value)
Phase: Redirection when an argument could not be "matched", i.e. either
there was no URL parameter, or the parameter could not be translated to
its original value (when translating a Realurl-encoded parameter back
into fx a UID). The feature is configured using the @route Redirect
annotation:

@route Redirect('NoMatch') $arg

This activates redirection when argument $arg is missing. Redirection
does work without being bound to an argument, although this does not
really make sense (as it would redirect every single time the page is
called…)

Additional configuration options are specified on the same line, such
as:

@route Redirect('NoMatch') Action('list')
Controller('AnotherController') $arg

Which will redirect to action "list" on AnotherController if $arg is
missing or could not be translated.

...and so on.

Supported commands and their parameters:

Action('lowercaseActionName')
Controller('UpperCaseName')
PluginName('UpperCaseName')
ExtensionName('UpperCaseName')
Prefix('tx_myext_pi') - required if using other commands
Status(302) - optional, has no default value
EntryLevel(-1) - optional, use to walk through the rootline (still
inactive)

This last argument, EntryLevel, is useful if your plugin always uses a
subpage to render a detail view and this always would redirect -1 level
up if the argument for the detail view is missing.

Note: current version does not yet redirect to pretty URLs and does not
consider EntryLevel.